refactor(ui): extract icon glyphs into morphir.ui.icons
